[Refactor/#29] 코드 구조 리펙토링 및 Api문서 작성
