+++ src/main/java/life/inha/icemarket/domain/User.java
@@ -10,7 +10,6 @@ import javax.validation.constraints.Email;
 import java.time.Instant;
 import java.util.List;
 
-
 @RequiredArgsConstructor
 @NoArgsConstructor
 @Getter
